Improved debugging messages and added memory utils.
Changed from using a lot of if statements and uart_puts to using debug_rc. This reduces some code bloat and just looks a lot nicer.

The addition of memory.rs is more complex. It is based off of https://github.com/brianwiddas/pi-baremetal/blob/master/memory.c. First issue is how to create a top level static variable. For the moment, I have a const pointer to the start of the page table that I will be using. Second issue is that the structs are not filled when using mem_v2p, might be an issue on implementation of this function, but I am not sure...