Testing, DMA, and better layout
Factored out all the tests in main.rs out to the src/bin/* directory to
reduce clutter.

Added DMA support through the dma.rs file. This allows for direct memory
access and variants of sdhci and vc gpu core have been created utilizing
DMA. This will be used more after the mmu has been implemented and I
have a heap to use.

VideoCore GPU code was made to make better use of traits and templating
so there isn't quite as much weird spaghetti code in postman.rs. Instead
there is a mailbox module with a vc.rs file to abstract out a mailbox
interface so that things won't be quite as much of a pain if I decide
that I want to add x86_64 support for krust.

UART is changed to have a trait interface and have the MiniUart as an
implementation of that trait. PL011 Uart will be implemented at some
point, but my #1 focus is getting a heap/allocator working so that a
filesystem can actually be used in krust.